added rect and square

+++ include/drawpp.hpp
@@ -17,13 +17,13 @@ void setTitle(const char* title);
 // Shapes 2D
 void arc(int x, int y, int width, int height, int start, int stop);
 void arc(int x, int y, int width, int height, int start, int stop, int mode);
-void circle(int x, int y, int radius);
-void ellipse(int x, int y, int width, int height);
+void circle(float x, float y, float radius);
+void ellipse(float x, float y, float width, float height);
 void line(int x1, int y1, int x2, int y2);
 void point(int x, int y);
 void quad(int x1, int y1, int x2, int y2, int x3, int y3, int x4, int y4);
-void rect(int x, int y, int w, int h);
-void square(int x, int y, int size);
+void rect(float x, float y, float w, float h);
+void square(float x, float y, float size);
 void triangle(int x1, int y1, int x2, int y2, int x3, int y3);
 
 // Shapes 3D
